+++ src/Material.h
@@ -9,10 +9,10 @@ using namespace std;
 class Material {
 public:
     Color color;
-    float ambient;
-    float diffuse;
-    float specular;
-    float shininess;
+    double ambient;
+    double diffuse;
+    double specular;
+    double shininess;
 
     Material() :color{Color(1,1,1)}, ambient{0.1}, diffuse{0.9}, specular{0.9}, shininess{200.0} { }
     bool operator== (const Material& rhs) const {
